- changed back to using no optimization for UML modules, and no -g2 - set inode operations in object create - s/WRITE/OBD_BRW_WRITE/ where applicable - added much of the infrastructure for page reads/writes - update inode->i_size on the MDS if extending the file - removed extra mntget() which hung the loop device - misc bugfixes - changed notify_change to i_op->setattr; otherwise we get the mtime and ctime   from the mds - set reply lengths (or your replies get dropped!)

+++ lustre/lib/obd_pack.c
@@ -245,7 +245,7 @@ void ost_unpack_niobuf(void **tmp, struct niobuf **nbp)
 
         *nbp = *tmp;
 
-        nb->addr = NTOH__u64((__u64)(unsigned long)nb->addr); 
+        nb->addr = NTOH__u64(nb->addr); 
         nb->offset = NTOH__u64(nb->offset); 
         nb->len = NTOH__u32(nb->len); 
         nb->flags = NTOH__u32(nb->flags);
